+++ src/Sulu/Bundle/PreviewBundle/Preview/Renderer/.preload.php
@@ -1,0 +1,5 @@
+<?php
+
+if (file_exists(__DIR__.'/../../../../../../../../../var/cache/preview/dev/App_KernelDevDebugContainer.preload.php')) {
+    require __DIR__.'/../../../../../../../../../var/cache/preview/dev/App_KernelDevDebugContainer.preload.php';
+}
